backend/requirements.txt
passlib
starlette
fastapi
sqlalchemy
pydantic
python-jose
python-dotenv